ING-424 extend memcachir to allow multiple herds
